split code into library and executable
