+++ ghc-dump-tree.cabal
@@ -13,6 +13,25 @@ build-type:          Simple
 cabal-version:       >=1.10
 tested-with:         GHC==7.4.2, GHC==7.8.3, GHC==7.10.0.20150310
 
+library
+  build-depends:       base                 >=4.5  && <4.9,
+                       aeson                >=0.7  && <0.9,
+                       bytestring           >=0.9  && <0.11,
+                       ghc                  >=7.4  && <7.12,
+                       optparse-applicative >=0.11 && <0.12,
+                       pretty               >=1.1  && <1.2,
+                       pretty-show          >=1.6  && <1.7,
+                       process              >=1.1  && <1.3,
+                       unordered-containers >=0.2  && <0.3,
+                       vector               >=0.10 && <0.11
+  hs-source-dirs:      src
+  default-language:    Haskell2010
+  ghc-options:         -Wall -fno-warn-orphans
+  other-extensions:    CPP
+  default-extensions:  RecordWildCards
+                       OverloadedStrings
+  exposed-modules:     Language.Haskell.GHC.DumpTree
+
 executable ghc-dump-tree
   main-is:             Main.hs
   build-depends:       base                 >=4.5  && <4.9,
@@ -20,10 +39,12 @@ executable ghc-dump-tree
                        bytestring           >=0.9  && <0.11,
                        ghc                  >=7.4  && <7.12,
                        optparse-applicative >=0.11 && <0.12,
+                       pretty               >=1.1  && <1.2,
                        pretty-show          >=1.6  && <1.7,
                        process              >=1.1  && <1.3,
                        unordered-containers >=0.2  && <0.3,
-                       vector               >=0.10 && <0.11
+                       vector               >=0.10 && <0.11,
+                       ghc-dump-tree
   hs-source-dirs:      src
   default-language:    Haskell2010
   ghc-options:         -Wall -fno-warn-orphans
